# swagger/organiser-app-server-v1.yaml
swagger: "2.0"
info:
  title: Organiser Application API
  description: Organiser application made for learning Java
  version: 1.0.0
host: orgap.pl
schemes:
  - http
  - https
consumes:
  - application/json
produces:
  - application/json

paths:
  /user:
    post:
      summary: Register user
      operationId: postUser
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - description: New user data
          in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/RegisterUserRequest'
      responses:
        200:
          description: Successful registration
          schema:
            $ref: '#/definitions/User'
        400:
          description: Client error
          schema:
            $ref: '#/definitions/Error'
        500:
          description: Server error
          schema:
            $ref: '#/definitions/Error'

  /user/{id}/meeting:
    post:
      summary: Add meeting.
      operationId: postMeeting
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - description: User id
          in: path
          name: id
          required: true
          type: integer
        - description: New meeting data
          in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/Meeting'
      responses:
        200:
          description: Successful operation
          schema:
            $ref: '#/definitions/Meeting'
        400:
          description: Client error
          schema:
            $ref: '#/definitions/Error'
        401:
          description: Unauthorized error
          schema:
            $ref: '#/definitions/Error'
        500:
          description: Server error
          schema:
            $ref: '#/definitions/Error'
    get:
      summary: Getting meetings list.
      operationId: getMeetings
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - description: User id
          in: path
          name: id
          required: true
          type: integer
        - description: Number of records to return
          in: query
          name: limit
          required: true
          type: integer
        - description: Initial index of records to return
          in: query
          name: offset
          required: true
          type: integer
        - description: Phrase to search
          in: query
          name: search
          required: false
          type: string
      responses:
        200:
          description: Successful operation - get meeting list
          schema:
            type: array
            items:
              $ref: '#/definitions/MeetingList'
        400:
          description: Client error
          schema:
            $ref: '#/definitions/Error'
        401:
          description: Unauthorized error
          schema:
            $ref: '#/definitions/Error'
        500:
          description: Server error
          schema:
            $ref: '#/definitions/Error'

  /user/{id}/meeting/{meeting_id}:
    get:
      summary: Getting single meeting
      operationId: getSingleMeeting
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - description: User id
          in: path
          name: id
          required: true
          type: integer
        - description: Id of a single meeting
          in: path
          name: meeting_id
          required: true
          type: integer
      responses:
        200:
          description: Successful operation - get meeting
          schema:
            $ref: '#/definitions/Meeting'
        400:
          description: Client error
          schema:
            $ref: '#/definitions/Error'
        500:
          description: Server error
          schema:
            $ref: '#/definitions/Error'

  /authenticate:
    post:
      summary: Authenticate user
      operationId: postAuthenticate
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - description: Credentials data.
          in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/AuthenticationRequest'
      responses:
        200:
          description: Successful login
          schema:
            $ref: '#/definitions/AuthenticationResponse'
        400:
          description: Client error
          schema:
            $ref: '#/definitions/Error'
        500:
          description: Server error
          schema:
            $ref: '#/definitions/Error'

definitions:
  Error:
    type: object
    properties:
      message:
        type: string
        description: Error message
    required:
      - message
    example:
      message: 'Apllication error'

  User:
    type: object
    properties:
      id:
        type: integer
        description: User id
      nick_name:
        type: string
        description: User nickname
      email:
        type: string
        description: User email address

  AuthenticationRequest:
    type: object
    properties:
      email:
        type: string
      password:
        type: string
    required:
      - email
      - password
    example:
      email: 'w.małachowski@java.pl'
      password: 'password'


  AuthenticationResponse:
    type: object
    properties:
      access_token:
        type: string
      expires_in:
        type: integer
      user_id:
        type: integer
      email:
        type: string
    required:
      - access_token
      - expires_in
      - user_id
      - email
    example:
      access_token: 'user_access_token'
      expires_in: 1800
      user_id: 1
      email: 'w.małachowski@java.pl'


  RegisterUserRequest:
    type: object
    properties:
      nick_name:
        type: string
        description: User nickname
      email:
        type: string
        description: User email address
      password:
        type: string
        description: User password
    required:
      - email
      - password
    example:
      nick_name: Mały
      email: 'w.małachowski@java.pl'
      password: 'password'


  Participant:
    type: object
    properties:
      id:
        type: integer
        description: Participant id
      first_name:
        type: string
        description: Participant first name
      last_name:
        type: string
        description: Participant last name
    required:
      - id
      - first_name
      - last_name
    example:
      id: 1
      first_name: Mateusz
      last_name: Kowalski

  Meeting:
    type: object
    properties:
      id:
        type: integer
        description: Meeting id
      name:
        type: string
        description: Meeting name
      date:
        type: string
        pattern: /([0-9]{4})-(?:[0-9]{2})-([0-9]{2})/
        description: Meeting date
      participants:
        type: array
        items:
          $ref: '#/definitions/Participant'
    required:
      - id
      - name
      - date
      - participants
    example:
      id: 1
      name: Alergolog
      date: '2021-04-12'
      participants:

  MeetingList:
    type: object
    properties:
      total:
        type: integer
        description: Movie list total items count
      date:
        type: array
        items:
          $ref: '#/definitions/Meeting'
        description: One pagination page data
    required:
      - total
      - data
    example:
      total: 99
      date:
        - id: 1
          name: Fryzjer
          date: 2021-04-12
        - id: 1
          name: Fryzjer
          date: 2021-04-12
        - id: 1
          name: Fryzjer
          date: 2021-04-12

  MeetingRequest:
    type: object
    properties:
      name:
        type: string
        description: Meeting name
      date:
        type: string
        pattern: /([0-9]{4})-(?:[0-9]{2})-([0-9]{2})/
        description: Meeting date
      participants:
        type: array
        items:
          $ref: '#/definitions/Participant'
    required:
      - id
      - name
      - date
    example:
      id: 1
      name: Fryzjer
      date: 2021-04-12